requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.6.0
Eel==0.14.0
future==0.18.2
geojson==2.5.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
io-chunks==1.0.2
multidict==5.1.0
numpy==1.20.1
Pillow==8.1.0
pycparser==2.20
pyowm==3.2.0
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PySocks==1.7.1
python-dotenv==0.15.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
simpy==4.0.1
simpy.io==0.2.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whichcraft==0.6.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.3.0